Providing more accurate data in mocked requests
